fix missing references, pull ReportGenerator.Reporting into ReportGenerator.Core, setup ReportGenerator.Core to be released as independant nuget package
