+++ src/Deployment/nuget/ReportGenerator.Core.nuspec
@@ -1,0 +1,28 @@
+<?xml version="1.0"?>
+<package xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
+  <metadata>
+    <id>ReportGenerator.Core</id>
+    <version>4.0.0.0</version>
+    <title />
+    <authors>Daniel Palme</authors>
+    <owners />
+    <licenseUrl>https://raw.githubusercontent.com/danielpalme/ReportGenerator/master/LICENSE.txt</licenseUrl>
+    <projectUrl>https://github.com/danielpalme/ReportGenerator</projectUrl>
+    <iconUrl>http://danielpalme.github.io/ReportGenerator/resources/logo.png</iconUrl>
+    <requireLicenseAcceptance>false</requireLicenseAcceptance>
+    <description>ReportGenerator converts XML reports generated by OpenCover, dotCover, Visual Studio, NCover, or Cobertura or into human readable reports in various formats. The reports do not only show the coverage quota, but also include the source code and visualize which line has been covered.</description>
+    <summary />
+    <language>en-US</language>
+    <tags>Code Coverage Reporting Testing TDD Utility</tags>
+    <dependencies>
+      <dependency id="SixLabors.ImageSharp.Drawing" version="1.0.0-beta0002" />
+    </dependencies>
+  </metadata>
+  <files>
+    <file src="..\..\target\bin\Release\ReportGenerator.Core.deps.json" target="lib\netstandard2.0\" />
+    <file src="..\..\target\bin\Release\ReportGenerator.Core.dll" target="lib\netstandard2.0\" />
+    <file src="..\..\target\bin\Release\LICENSE.txt" target="LICENSE.txt" />
+    <file src="..\..\target\bin\Release\Readme.txt" target="Readme.txt" />
+  </files>
+</package>
+
