+++ src/Deployment/nuget/ReportGenerator.nuspec
@@ -17,21 +17,26 @@
     <packageTypes>
       <packageType name="DotnetCliTool" />
     </packageTypes>
+    <dependencies>
+      <dependency id="SixLabors.ImageSharp.Drawing" version="1.0.0-beta0002" />
+    </dependencies>
   </metadata>
   <files>
-  <file src="..\..\target\bin\Release\dotnet-ReportGenerator.deps.json" target="lib\netcoreapp2.0\" />
+    <!-- DotNet CLI -->
+    <file src="..\..\target\bin\Release\dotnet-ReportGenerator.deps.json" target="lib\netcoreapp2.0\" />
     <file src="..\..\target\bin\Release\dotnet-ReportGenerator.runtimeconfig.json" target="lib\netcoreapp2.0\" />
     <file src="..\..\target\bin\Release\dotnet-ReportGenerator.dll" target="lib\netcoreapp2.0\" />
     <file src="..\..\target\bin\Release\ReportGenerator.Core.deps.json" target="lib\netcoreapp2.0\" />
     <file src="..\..\target\bin\Release\ReportGenerator.Core.dll" target="lib\netcoreapp2.0\" />
-    <file src="..\..\target\bin\Release\ReportGenerator.Reporting.deps.json" target="lib\netcoreapp2.0\" />
-    <file src="..\..\target\bin\Release\ReportGenerator.Reporting.dll" target="lib\netcoreapp2.0\" />
+    <!-- Legacy EXE -->
+    <file src="..\..\target\bin\Release\Microsoft.Win32.Primitives.dll" target="tools" />
+    <file src="..\..\target\bin\Release\netstandard.dll" target="tools" />
     <file src="..\..\target\bin\Release\ReportGenerator.exe" target="tools" />
     <file src="..\..\target\bin\Release\ReportGenerator.exe.config" target="tools" />
     <file src="..\..\target\bin\Release\ReportGenerator.Core.deps.json" target="tools" />
     <file src="..\..\target\bin\Release\ReportGenerator.Core.dll" target="tools" />
-    <file src="..\..\target\bin\Release\ReportGenerator.Reporting.deps.json" target="tools" />
-    <file src="..\..\target\bin\Release\ReportGenerator.Reporting.dll" target="tools" />
+    <file src="..\..\target\bin\Release\SixLabors.*.dll" target="tools" />
+    <file src="..\..\target\bin\Release\System.*.dll" target="tools" />
     <file src="..\..\target\bin\Release\LICENSE.txt" target="LICENSE.txt" />
     <file src="..\..\target\bin\Release\Readme.txt" target="Readme.txt" />
   </files>
