+++ src/ReportGenerator.Core/Reporting/Builders/HtmlChartReportBuilder.cs
@@ -3,10 +3,10 @@ using System.Collections.Generic;
 using System.Linq;
 using Palmmedia.ReportGenerator.Core.Parser.Analysis;
 using Palmmedia.ReportGenerator.Core.Properties;
-using Palmmedia.ReportGenerator.Reporting.CodeAnalysis;
-using Palmmedia.ReportGenerator.Reporting.Rendering;
+using Palmmedia.ReportGenerator.Core.Reporting.CodeAnalysis;
+using Palmmedia.ReportGenerator.Core.Reporting.Rendering;
 
-namespace Palmmedia.ReportGenerator.Reporting
+namespace Palmmedia.ReportGenerator.Core.Reporting.Builders
 {
     /// <summary>
     /// Creates HTML with chart component only (no reports for classes are generated).
