+++ src/ReportGenerator.Core/Reporting/Builders/HtmlReportBuilder.cs
@@ -1,22 +1,27 @@
 ﻿using System.Collections.Generic;
 using System.Text;
 using Palmmedia.ReportGenerator.Core.Parser.Analysis;
-using Palmmedia.ReportGenerator.Reporting.Rendering;
+using Palmmedia.ReportGenerator.Core.Reporting.Rendering;
 
-namespace Palmmedia.ReportGenerator.Reporting
+namespace Palmmedia.ReportGenerator.Core.Reporting.Builders
 {
     /// <summary>
-    /// Creates summary report in HTML format (no reports for classes are generated).
+    /// Creates report in HTML format.
     /// </summary>
-    public class HtmlSummaryReportBuilder : ReportBuilderBase
+    public class HtmlReportBuilder : ReportBuilderBase
     {
+        /// <summary>
+        /// Contains report specific JavaScript content.
+        /// </summary>
+        private readonly StringBuilder javaScriptContent = new StringBuilder();
+
         /// <summary>
         /// Gets the report type.
         /// </summary>
         /// <value>
         /// The report format.
         /// </value>
-        public override string ReportType => "HtmlSummary";
+        public override string ReportType => "Html";
 
         /// <summary>
         /// Gets a value indicating whether class reports can be generated in parallel.
@@ -30,6 +35,10 @@ namespace Palmmedia.ReportGenerator.Reporting
         /// <param name="fileAnalyses">The file analyses that correspond to the class.</param>
         public override void CreateClassReport(Class @class, IEnumerable<FileAnalysis> fileAnalyses)
         {
+            using (var renderer = new HtmlRenderer(false, false, this.javaScriptContent))
+            {
+                this.CreateClassReport(renderer, @class, fileAnalyses);
+            }
         }
 
         /// <summary>
@@ -38,7 +47,7 @@ namespace Palmmedia.ReportGenerator.Reporting
         /// <param name="summaryResult">The summary result.</param>
         public override void CreateSummaryReport(SummaryResult summaryResult)
         {
-            using (var renderer = new HtmlRenderer(true, true, new StringBuilder()))
+            using (var renderer = new HtmlRenderer(false, false, this.javaScriptContent))
             {
                 this.CreateSummaryReport(renderer, summaryResult);
             }
