+++ src/ReportGenerator.Core/Reporting/Builders/LatexSummaryReportBuilder.cs
@@ -1,27 +1,21 @@
 ﻿using System.Collections.Generic;
-using System.Text;
 using Palmmedia.ReportGenerator.Core.Parser.Analysis;
-using Palmmedia.ReportGenerator.Reporting.Rendering;
+using Palmmedia.ReportGenerator.Core.Reporting.Rendering;
 
-namespace Palmmedia.ReportGenerator.Reporting
+namespace Palmmedia.ReportGenerator.Core.Reporting.Builders
 {
     /// <summary>
-    /// Creates report in HTML format.
+    /// Creates summary report in Latex format (no reports for classes are generated).
     /// </summary>
-    public class HtmlInlineCssAndJavaScriptReportBuilder : ReportBuilderBase
+    public class LatexSummaryReportBuilder : ReportBuilderBase
     {
-        /// <summary>
-        /// Contains report specific JavaScript content.
-        /// </summary>
-        private readonly StringBuilder javaScriptContent = new StringBuilder();
-
         /// <summary>
         /// Gets the report type.
         /// </summary>
         /// <value>
         /// The report format.
         /// </value>
-        public override string ReportType => "HtmlInline";
+        public override string ReportType => "LatexSummary";
 
         /// <summary>
         /// Gets a value indicating whether class reports can be generated in parallel.
@@ -35,10 +29,6 @@ namespace Palmmedia.ReportGenerator.Reporting
         /// <param name="fileAnalyses">The file analyses that correspond to the class.</param>
         public override void CreateClassReport(Class @class, IEnumerable<FileAnalysis> fileAnalyses)
         {
-            using (var renderer = new HtmlRenderer(false, true, this.javaScriptContent))
-            {
-                this.CreateClassReport(renderer, @class, fileAnalyses);
-            }
         }
 
         /// <summary>
@@ -47,7 +37,7 @@ namespace Palmmedia.ReportGenerator.Reporting
         /// <param name="summaryResult">The summary result.</param>
         public override void CreateSummaryReport(SummaryResult summaryResult)
         {
-            using (var renderer = new HtmlRenderer(false, true, this.javaScriptContent))
+            using (var renderer = new LatexRenderer())
             {
                 this.CreateSummaryReport(renderer, summaryResult);
             }
