+++ src/ReportGenerator.Core/Reporting/CodeAnalysis/RiskHotspot.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using Palmmedia.ReportGenerator.Core.Parser.Analysis;
 
-namespace Palmmedia.ReportGenerator.Reporting.CodeAnalysis
+namespace Palmmedia.ReportGenerator.Core.Reporting.CodeAnalysis
 {
     /// <summary>
     /// Represents a risk hotspot.
