+++ src/ReportGenerator.Core/Reporting/Rendering/IReportRenderer.cs
@@ -1,8 +1,8 @@
 ﻿using System.Collections.Generic;
 using Palmmedia.ReportGenerator.Core.Parser.Analysis;
-using Palmmedia.ReportGenerator.Reporting.CodeAnalysis;
+using Palmmedia.ReportGenerator.Core.Reporting.CodeAnalysis;
 
-namespace Palmmedia.ReportGenerator.Reporting.Rendering
+namespace Palmmedia.ReportGenerator.Core.Reporting.Rendering
 {
     /// <summary>
     /// Interface for report renderers.
