+++ src/ReportGenerator.Core/Reporting/Rendering/LatexRenderer.cs
@@ -7,9 +7,9 @@ using System.Text;
 using Palmmedia.ReportGenerator.Core.Parser.Analysis;
 using Palmmedia.ReportGenerator.Core.Properties;
 using Palmmedia.ReportGenerator.Core.Reporting;
-using Palmmedia.ReportGenerator.Reporting.CodeAnalysis;
+using Palmmedia.ReportGenerator.Core.Reporting.CodeAnalysis;
 
-namespace Palmmedia.ReportGenerator.Reporting.Rendering
+namespace Palmmedia.ReportGenerator.Core.Reporting.Rendering
 {
     /// <summary>
     /// Latex report renderer.
