+++ src/ReportGenerator.Core/Reporting/Rendering/resources/coverageTableReactComponents.js
