+++ src/ReportGenerator.Core/Reporting/Rendering/resources/custom.css
