+++ src/ReportGenerator.Core/Reporting/Rendering/resources/customAngularApp.js
