+++ src/ReportGenerator.Core/Reporting/Rendering/resources/riskHotspotTableReactComponents.js
