--- src/ReportGenerator.Reporting/ReportGenerator.Reporting.csproj
@@ -1,56 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <RootNamespace>Palmmedia.ReportGenerator.Reporting</RootNamespace>
-    <AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
-    <CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
-    <AssemblyVersion>4.0.0.0</AssemblyVersion>
-    <FileVersion>4.0.0.0</FileVersion>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-    <OutputPath>..\target\bin\Debug\</OutputPath>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
-    <OutputPath>..\target\bin\Release\</OutputPath>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <None Remove="Rendering\resources\angular.min.js" />
-    <None Remove="Rendering\resources\chartist.min.css" />
-    <None Remove="Rendering\resources\chartist.min.js" />
-    <None Remove="Rendering\resources\coverageTableReactComponents.js" />
-    <None Remove="Rendering\resources\custom.css" />
-    <None Remove="Rendering\resources\customAngularApp.js" />
-    <None Remove="Rendering\resources\jquery-1.11.2.min.js" />
-    <None Remove="Rendering\resources\matchMedia.js" />
-    <None Remove="Rendering\resources\react.min.js" />
-    <None Remove="Rendering\resources\react.modified.min.js" />
-    <None Remove="Rendering\resources\riskHotspotTableReactComponents.js" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <EmbeddedResource Include="Rendering\resources\angular.min.js" />
-    <EmbeddedResource Include="Rendering\resources\chartist.min.css" />
-    <EmbeddedResource Include="Rendering\resources\chartist.min.js" />
-    <EmbeddedResource Include="Rendering\resources\coverageTableReactComponents.js" />
-    <EmbeddedResource Include="Rendering\resources\custom.css" />
-    <EmbeddedResource Include="Rendering\resources\customAngularApp.js" />
-    <EmbeddedResource Include="Rendering\resources\jquery-1.11.2.min.js" />
-    <EmbeddedResource Include="Rendering\resources\matchMedia.js" />
-    <EmbeddedResource Include="Rendering\resources\react.min.js" />
-    <EmbeddedResource Include="Rendering\resources\react.modified.min.js" />
-    <EmbeddedResource Include="Rendering\resources\riskHotspotTableReactComponents.js" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="SixLabors.ImageSharp.Drawing" Version="1.0.0-beta0002" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\ReportGenerator.Core\ReportGenerator.Core.csproj" />
-  </ItemGroup>
-
-</Project>
